feat(api): Fetch RSVP user IDs with activities
Modify the `getActivities` function to join the `activities` table with the `activity_rsvps` table. This change ensures that the list of user IDs who have RSVP'd to an activity is correctly fetched and included in the activity data. The `rsvpUserIds` field in the returned `Activity` object is now populated from the `activity_rsvps` relationship.